+++ src/RcppExports.cpp
@@ -41,8 +41,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // genome_to_libsvm
-bool genome_to_libsvm(const CharacterVector& x, const CharacterVector& target_path, const CharacterVector& label, int k, bool canonical, bool squeeze);
-RcppExport SEXP _MIC_genome_to_libsvm(SEXP xSEXP, SEXP target_pathSEXP, SEXP labelSEXP, SEXP kSEXP, SEXP canonicalSEXP, SEXP squeezeSEXP) {
+bool genome_to_libsvm(const CharacterVector& x, const CharacterVector& target_path, const CharacterVector& label, int k, bool canonical, bool squeeze, bool overwrite);
+RcppExport SEXP _MIC_genome_to_libsvm(SEXP xSEXP, SEXP target_pathSEXP, SEXP labelSEXP, SEXP kSEXP, SEXP canonicalSEXP, SEXP squeezeSEXP, SEXP overwriteSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -52,7 +52,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type k(kSEXP);
     Rcpp::traits::input_parameter< bool >::type canonical(canonicalSEXP);
     Rcpp::traits::input_parameter< bool >::type squeeze(squeezeSEXP);
-    rcpp_result_gen = Rcpp::wrap(genome_to_libsvm(x, target_path, label, k, canonical, squeeze));
+    Rcpp::traits::input_parameter< bool >::type overwrite(overwriteSEXP);
+    rcpp_result_gen = Rcpp::wrap(genome_to_libsvm(x, target_path, label, k, canonical, squeeze, overwrite));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -108,7 +109,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_MIC_reverse_complement", (DL_FUNC) &_MIC_reverse_complement, 1},
     {"_MIC_kmers", (DL_FUNC) &_MIC_kmers, 9},
-    {"_MIC_genome_to_libsvm", (DL_FUNC) &_MIC_genome_to_libsvm, 6},
+    {"_MIC_genome_to_libsvm", (DL_FUNC) &_MIC_genome_to_libsvm, 7},
     {"_MIC_squeezed_mers", (DL_FUNC) &_MIC_squeezed_mers, 1},
     {"_MIC_unsqueezed_mers", (DL_FUNC) &_MIC_unsqueezed_mers, 1},
     {"_MIC_squeezed_index_to_str", (DL_FUNC) &_MIC_squeezed_index_to_str, 3},